Move branch.ts from lib/ -> client/
