+++ packages/api/src/client/branch.ts
@@ -2,9 +2,9 @@ import { valid } from '@urbit/aura';
 
 import { getPostInfoFromWer } from '../client/harkApi';
 import { createDevLogger } from '../lib/logger';
+import { normalizeUrbitColor } from '../lib/utils';
 import { getConstants } from '../types/constants';
 import { preSig } from '../urbit';
-import { normalizeUrbitColor } from './utils';
 
 const logger = createDevLogger('branch', false);
