+++ packages/api/src/lib/deeplinks.ts
@@ -1,8 +1,8 @@
+import { AppInvite, getBranchLinkMeta, isLureMeta } from '../client/branch';
 import { createDevLogger } from '../lib/logger';
 import { getConstants } from '../types/constants';
 import type { ContentReference } from '../types/references';
 import { citeToPath } from '../urbit';
-import { AppInvite, getBranchLinkMeta, isLureMeta } from './branch';
 import { normalizeUrbitColor } from './utils';
 
 const logger = createDevLogger('deeplinks', false);
